requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
pymemcache==4.0.0
PyYAML==6.0.1

